refactor(service tiles): duplicated code
- extracted duplicated code from company and service screens
