+++ src/components/screens/index.js
@@ -7,4 +7,5 @@ export * from './OperatingCompany';
 export * from './PointOfInterest';
 export * from './PriceCard';
 export * from './Service';
+export * from './ServiceTiles';
 export * from './Tour';
